requirements.txt
mcp
langchain-google-genai
langchain
langchain-community
yfinance
python-dotenv
fastmcp
